refactor: output 관련 내용을 ViewController로부터 OutputController로 분리"
